src/main/java/com/classroom/assignment/controller/api/Step2ApiController.java
package com.classroom.assignment.controller.api;

import java.util.concurrent.atomic.AtomicLong;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import com.classroom.assignment.model.response.Greeting;

@RestController
@RequestMapping("/api/step2")
public class Step2ApiController {

  private static final String template = "Hello, %s!";
  private final AtomicLong counter = new AtomicLong();

  @GetMapping
  public String index() {
    return "index";
  }

  @GetMapping("/greeting")
  public Greeting greeting(@RequestParam(value = "name", defaultValue = "World") final String name) {
    return new Greeting(1l, String.format(template, name));
  }
}
